mise en place adresse et ajout|infos

+++ pje-android/ProjetContact/app/src/main/java/com/example/projetcontact/MainActivity.java
@@ -22,18 +22,14 @@ import android.widget.Toast;
 import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
-
-    public static final String EXTRA_NOM = "nom";
-    public static final String EXTRA_AGE = "age";
-    public static final String EXTRA_PRENOM = "prenom";
-
     public static ContactViewModel mContactViewModel;
 
     public static final int NEW_CONTACT_ACTIVITY_REQUEST_CODE = 1;
-    /*
+    public static Contact newContact;
+
     public static final int UPDATE_ACTIVITY_REQUEST_CODE = 2;
     public static Contact updateContact;
-    */
+
     public static final int DELETE_ACTIVITY_REQUEST_CODE = 3;
     public static Contact removeContact;
 
@@ -94,17 +90,10 @@ public class MainActivity extends AppCompatActivity {
         super.onActivityResult(requestCode, resultCode, data);
 
         if (requestCode == NEW_CONTACT_ACTIVITY_REQUEST_CODE && resultCode == RESULT_OK) {
-            String n = data.getStringExtra(EXTRA_NOM);
-            String p = data.getStringExtra(EXTRA_PRENOM);
-            String a = data.getStringExtra(EXTRA_AGE);
-            Contact contact = new Contact();
-            contact.setNom(n);
-            contact.setPrenom(p);
-            contact.setAge(a);
-            mContactViewModel.insert(contact);
-        } /* if (requestCode == UPDATE_ACTIVITY_REQUEST_CODE && resultCode == RESULT_OK){
+            mContactViewModel.insert(newContact);
+        }  if (requestCode == UPDATE_ACTIVITY_REQUEST_CODE && resultCode == RESULT_OK){
             mContactViewModel.updateContact(updateContact);
-        } */ if (requestCode == DELETE_ACTIVITY_REQUEST_CODE && resultCode == RESULT_OK){
+        }  if (requestCode == DELETE_ACTIVITY_REQUEST_CODE && resultCode == RESULT_OK){
             mContactViewModel.delete(removeContact);
         } else {
             Toast.makeText(
@@ -113,13 +102,13 @@ public class MainActivity extends AppCompatActivity {
                     Toast.LENGTH_LONG).show();
         }
     }
-/*
+
     public void infosContact(Contact contact){
         Intent intent = new Intent(this,InfoContactActivity.class);
         updateContact = contact;
         startActivityForResult(intent, UPDATE_ACTIVITY_REQUEST_CODE);
     }
-*/
+
     public void removeContact(Contact contact){
         Intent intent = new Intent(this, DeleteContactActivity.class);
         removeContact = contact;
